Update README.md
Solving the Quadratic Equation

Write a program that solves a quadratic equation to find its roots.

The roots of a quadratic equation
ax^2 + bx + c = 0
(where a is not zero) are given by the formula
(−b ± sqrt(b^2 − 4ac)) / 2a

The value of the discriminant (b^2 − 4ac) determines the nature of roots. If the value of the discriminant is zero, then the equation has a single real root. If the value of the discriminant is positive then the equation has two real roots. If the value of the discriminant is negative, then the equation has two complex roots. For this assignment, you can assume the discriminant is non-negative.

The program takes values of a, b, and c as input and outputs the roots. You do not need to output complex roots.